backend/requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
click==8.2.1
fastapi==0.104.1
GeoAlchemy2==0.14.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.2
packaging==25.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
shapely
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.14.1
uvicorn==0.24.0
